Update Dice Game.py
This is my very first project using Python. Its a dice game where there are 2 players, each have up to throws of a pair of dices. The one with the larger sum wins. Each time you throw the dices, that result is your new score, erasing the previous one, and you can stop at any throw after the first one.

It would be nice to get some feedback on my code, I know there might be lot of ways to improve and make it shorter.